requirements.txt
python-telegram-bot
tflite-runtime==2.10.0
numpy
Flask
matplotlib